Added and configured the wysiwyg_filter module.
This module and configuration was used in the websolutions profile. I have
copied over the basic configuration from there, except where the websolutions
profile allowed all class names, I have limited the allowed class names to
bootstrap classes.

The list of allowed tags may need to be expanded for AODA sites, but this
is a starting point.