Move navbar button after brand in global navbar.
Moved the menu expand/collapse button after the "brand" (U of G logo)
in the global (red) navbar menu. Previously the button was before
the brand in document order (as it is in the Bootstrap navbar
documentation). This looks okay, but it means that when you navigating
with the keyboard the "new" menu items that are added when you click
the expand button do not immediately follow the button (in document
order) which violates WCAG SCR26
(http://www.w3.org/TR/2014/NOTE-WCAG20-TECHS-20140916/SCR26.html)

The new ordering looks the same but is more accessible.

Fixes issue #125.